(map_sequence) Combine playerphase.cpp and mind.cpp with ProcSeq, ProcCP, and assorted others
